separate out classes to meaningful files
fix tests
